[qacode] renamed test suites
