+++ tests/functionals/TestInfoBotMultiple.py
@@ -8,18 +8,19 @@ from qacode.core.loggers.LoggerManager import LoggerManager
 from qacode.core.testing.TestInfoBot import TestInfoBot
 from qacode.core.utils.Utils import settings
 
+
 LOGGER_MANAGER = LoggerManager()
 SETTINGS = settings()
-BOT = TestInfoBot.bot_open(
-    test_config=SETTINGS,
-    logger_manager=LOGGER_MANAGER)
 
-class TestTestInfoBotUnique(TestInfoBot):
+
+class TestInfoBotMultiple(TestInfoBot):
     """Tests for class TestInfoBot"""
 
-    def __init__(self, method_name="TestTestInfoBotUnique"):
-        super(TestTestInfoBotUnique, self).__init__(
-            method_name, bot=BOT)
+    def __init__(self, method_name="TestTestInfoBot"):
+        super(TestInfoBotMultiple, self).__init__(
+            method_name,
+            logger_manager=LOGGER_MANAGER,
+            test_config=SETTINGS)
 
     def test_001_inheritance(self):
         """Testcase: test_001_inheritance"""
@@ -37,7 +38,3 @@ class TestTestInfoBotUnique(TestInfoBot):
         """Testcase: test_003_bot_instance"""
         self.log.info("Check bot instance")
         self.assertIsInstance(self.bot, BotBase)
-
-    @classmethod
-    def tearDownClass(cls):
-        TestInfoBot.bot_close(BOT)
