+++ tests/functionals/TestInfoBotUnique.py
@@ -6,18 +6,27 @@ import unittest
 from qacode.core.bots.BotBase import BotBase
 from qacode.core.loggers.LoggerManager import LoggerManager
 from qacode.core.testing.TestInfoBot import TestInfoBot
+from qacode.core.utils.Utils import settings
 
 LOGGER_MANAGER = LoggerManager()
+SETTINGS = settings()
+BOT = None
 
-
-class TestTestInfoBot(TestInfoBot):
+class TestInfoBotUnique(TestInfoBot):
     """Tests for class TestInfoBot"""
 
-    def __init__(self, method_name="TestTestInfoBot"):
-        super(TestTestInfoBot, self).__init__(
-            method_name,
-            logger_manager=LOGGER_MANAGER,
-            test_config=None)
+    @classmethod
+    def setUpClass(cls):
+        global BOT
+        BOT = TestInfoBot.bot_open(SETTINGS, LOGGER_MANAGER)
+    @classmethod
+    def tearDownClass(cls):
+        global BOT
+        TestInfoBot.bot_close(BOT)
+
+    def __init__(self, method_name="TestTestInfoBotUnique"):
+        super(TestInfoBotUnique, self).__init__(
+            method_name, bot=BOT)
 
     def test_001_inheritance(self):
         """Testcase: test_001_inheritance"""
